+++ RELEASE-NOTES.md
@@ -1,14 +1,31 @@
 Push Notification module for PlayFramework
 =========================
 
+## 1.1 - 2018-03-23
+- Support all ebean server and datasource config flags in custom application.conf properties.
+
+- Modify DAO to replace deprecated ebean method calls.
+
+- Fix EBean provider lazyloading instances.
+
+- Update dependencies (Play, ebean, Gson).
+
+- Improvements to the lifecycle manager to more gracefully startup and shutdown TaskQueue and datastore.
+
+- Miscellaneous bugfixes. 
+
+  ​
+
+
 ## 1.0.2 - 2017-08-23
-
  - Improvements to default database configuration
- 
-## 1.0.1 - 2017-08-22
 
+   ​
+
+## 1.0.1 - 2017-08-22
  - Sample application and documentation
 
-## 1.0.0 - 2017-08-21
+   ​
 
+## 1.0.0 - 2017-08-21
  - Initial release